requirements.txt
absl-py==2.2.2
asttokens==3.0.0
chex==0.1.89
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
etils==1.12.2
exceptiongroup==1.2.2
executing==2.2.0
flax==0.10.2
fonttools==4.57.0
fsspec==2025.3.2
humanize==4.12.2
importlib_resources==6.5.2
ipython==8.35.0
jax==0.4.35
jax-cuda12-pjrt==0.4.35
jax-cuda12-plugin==0.4.35
jaxlib==0.4.34
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.2.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.83
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.93
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
tensorstore==0.1.73
threadpoolctl==3.6.0
toolz==1.0.0
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
wcwidth==0.2.13
zipp==3.21.0
